ark integration: loads transactions tab
