+++ src/renderer/account/actions/transactionHistoryActions.js
@@ -1,87 +1,22 @@
-import fetch from 'node-fetch';
-import uuid from 'uuid/v4';
 import { createActions } from 'spunky';
-import { omit } from 'lodash';
 
-import { api } from '@cityofzion/neon-js';
+import getNEOTransactionHistory from './NEO/getNEOTransactionHistory';
+import getARKTransactionHistory from './ARK/getARKTransactionHistory';
 
-import { NEO, GAS, ARK, ASSETS } from 'shared/values/assets';
-import getTokens from 'shared/util/getTokens';
-
-const TX_TYPES = {
-  SEND: 'Send',
-  RECEIVE: 'Receive',
-  CLAIM: 'Claim',
-  INVOCATION: 'Invocation'
-};
-
-function parseAbstractData(data, currentUserAddress, tokens) {
-  const parsedTxType = (abstract) => {
-    if (abstract.address_to === currentUserAddress && abstract.address_from !== 'claim') {
-      return TX_TYPES.RECEIVE;
-    }
-    if (abstract.address_from === 'claim') return TX_TYPES.CLAIM;
-    return TX_TYPES.SEND;
-  };
-
-  const parsedAsset = (abstract) => {
-    const tokensResult = tokens.find(({ scriptHash }) => scriptHash === abstract.asset);
-    if (tokensResult) return tokensResult;
-    if (abstract.asset === NEO || abstract.asset === GAS || abstract.asset === ARK) {
-      return { ...ASSETS[abstract.asset], scriptHash: abstract.asset };
-    }
-    return { symbol: '' };
-  };
-
-  const parsedTo = (abstract) => {
-    if (abstract.address_to === 'fees') return 'NETWORK FEES';
-    if (abstract.address_to === 'mint') return 'MINT TOKENS';
-    return abstract.address_to;
-  };
-
-  const parsedFrom = (abstract) => {
-    if (abstract.address_from === 'mint') return 'MINT TOKENS';
-    return abstract.address_from;
-  };
-
-  return data.map((abstract) => {
-    const asset = parsedAsset(abstract);
-    const type = parsedTxType(abstract);
-    const summary = {
-      to: parsedTo(abstract),
-      isNetworkFee: abstract.address_to === 'fees',
-      from: parsedFrom(abstract),
-      txId: abstract.txid,
-      time: abstract.time,
-      amount: abstract.amount,
-      asset,
-      label: abstract.address_to === currentUserAddress ? 'IN' : 'OUT',
-      type,
-      id: uuid()
-    };
-
-    return summary;
-  });
-}
-
-async function getTransactionHistory({ net, address, previousCall = {} }) {
-  const { page_number: pageNumber = 0, total_pages: totalPages = 1, entries = [] } = previousCall;
-  const pageCount = pageNumber + 1;
-  if (pageCount > totalPages) return previousCall;
-
-  const tokens = await getTokens();
-  const endpoint = api.neoscan.getAPIEndpoint(net);
-
-  const data = await fetch(`${endpoint}/v1/get_address_abstracts/${address}/${pageCount}`);
-  const response = await data.json();
-  return {
-    ...omit(response, 'entries'),
-    entries: [...entries, ...parseAbstractData(response.entries, address, tokens)]
-  };
+async function getTransactionHistory({ net, coinType, address, previousCall = {} }) {
+  if (coinType === 888) {
+    // NEO
+    return getNEOTransactionHistory({ net, address, previousCall });
+  } else if (coinType === 111) {
+    // ARK
+    return getARKTransactionHistory({ address, previousCall });
+  } else {
+    throw new Error('coinType is invalid');
+  }
 }
 
 export const ID = 'transaction_history';
 
-export default createActions(ID, ({ net, address, previousCall }) => async () => {
-  return getTransactionHistory({ net, address, previousCall });
+export default createActions(ID, ({ net, coinType, address, previousCall }) => async () => {
+  return getTransactionHistory({ net, coinType, address, previousCall });
 });
