+++ src/renderer/shared/util/ARK/getArkBalance.js
@@ -5,7 +5,7 @@ export default async function getArkBalance({ address }) {
   const response = await fetch(url);
   const { data } = await response.json();
   const { balance } = data;
-  const number = (parseInt(balance, 10) / 1e8).toString();
+  const number = Number((parseFloat(balance, 10) / 1e8).toFixed(8)).toString();
   return {
     ARK: { name: 'ARK', symbol: 'ARK', scriptHash: 'ARK', decimals: 8, balance: number }
   };
